requirements.txt
streamlit>=1.31.0
sqlalchemy>=2.0.0
google-genai>=0.3.0
pytest>=7.4.0
pydantic>=1.10,<2.0
